# feat(pack): Performance Optimization for Large Repositories