--- a/src/core/file/fileProcess.ts
+++ b/src/core/file/fileProcess.ts
@@ -1,23 +1,13 @@
 import pc from 'picocolors';
-import { Piscina } from 'piscina';
 import type { RepomixConfigMerged } from '../../config/configSchema.js';
 import { logger } from '../../shared/logger.js';
-import { getWorkerThreadCount } from '../../shared/processConcurrency.js';
+import { initPiscina } from '../../shared/processConcurrency.js';
 import type { RepomixProgressCallback } from '../../shared/types.js';
 import type { ProcessedFile, RawFile } from './fileTypes.js';
 import type { FileProcessTask } from './workers/fileProcessWorker.js';
 
 const initTaskRunner = (numOfTasks: number) => {
-  const { minThreads, maxThreads } = getWorkerThreadCount(numOfTasks);
-  logger.trace(`Initializing worker pool with min=${minThreads}, max=${maxThreads} threads`);
-
-  const pool = new Piscina({
-    filename: new URL('./workers/fileProcessWorker.js', import.meta.url).href,
-    minThreads,
-    maxThreads,
-    idleTimeout: 5000,
-  });
-
+  const pool = initPiscina(numOfTasks, new URL('./workers/fileProcessWorker.js', import.meta.url).href);
   return (task: FileProcessTask) => pool.run(task);
 };