--- a/src/core/file/workers/types.ts
+++ b/src/core/file/workers/types.ts
@@ -1,0 +1,16 @@
+/**
+ * Task definition for file reading worker
+ */
+export interface ReadFileTask {
+  filePath: string;
+  rootDir: string;
+}
+
+/**
+ * Configuration for worker thread pool
+ */
+export interface WorkerPoolConfig {
+  minThreads?: number;
+  maxThreads?: number;
+  idleTimeout?: number;
+}