## Performance Improvement

### yamadashy/repomix
- Before: `868.73 millis (usr: 1.11 secs, sys: 0.14 secs)`
- After: `671.26 millis (usr: 1.42 secs, sys: 0.22 secs)`

No significant change

### facebook/react

- Before: `123.31 secs (usr: 118.64s, sys: 1.60s)`
- After: `7.10 secs (usr: 20.01s, sys: 2.39s)`

17x faster

### vercel/next.js
- Before: `17.85 mins (usr: 16.66 mins, sys: 0.18 mins)`
- After: `17.27 secs (usr: 52.93 secs, sys: 7.11 secs)`

58x faster

## Changes
- Replace `p-map` with Piscina worker threads for parallel processing
- Implement dedicated workers for file collection, processing, and metrics
- Optimize file search with single worker implementation
- Add proper resource cleanup for worker pools
- Move file manipulation logic to worker threads

## Checklist

- [x] Run `npm run test`
- [x] Run `npm run lint`